move getSpellDamage and parseMaterialComponents to parser.ts
