+++ src/dnd/spells.ts
@@ -6,8 +6,10 @@ import {
     parseDescriptions,
     parseDurationTime,
     parseImageUrl,
+    parseMaterialComponents,
     parseRange,
     parseReprint,
+    parseSpellDamage,
     parseSpellLevel,
     parseSpellSchool,
     ReprintData,
@@ -20,7 +22,7 @@ interface Caster {
     source: string;
 }
 
-interface SpellDamage {
+export interface SpellDamage {
     type: 'level' | 'upcast';
     scaling: { [key: number]: string };
 }
@@ -102,84 +104,6 @@ function getCasters(spell: any, sources: any[]): any[] {
     return casters;
 }
 
-function getSpellDamage(spell: any): SpellDamage[] | null {
-    if (spell.scalingLevelDice) {
-        const results: SpellDamage[] = [];
-        const scalingList = Array.isArray(spell.scalingLevelDice) ? spell.scalingLevelDice : [spell.scalingLevelDice];
-
-        for (const scaleObj of scalingList) {
-            if (scaleObj && scaleObj.scaling) {
-                const scaling: { [key: string]: string } = {};
-                for (const [key, value] of Object.entries(scaleObj.scaling)) {
-                    scaling[String(key)] = String(value);
-                }
-                results.push({ type: 'level', scaling });
-            }
-        }
-
-        return results;
-    }
-
-    const raw = [...(spell.entries || [])];
-    if (spell.entriesHigherLevel) {
-        raw.push(...spell.entriesHigherLevel);
-    }
-
-    const entriesText: string[] = [];
-    function pushEntryTexts(entries: any) {
-        for (const entry of entries) {
-            if (typeof entry === 'string') {
-                entriesText.push(entry);
-            } else if (entry && entry.entries) {
-                pushEntryTexts(entry.entries);
-            }
-        }
-    }
-    pushEntryTexts(raw);
-
-    const scaleDamageRegex = /{@scaledamage\s+([^}]+)}/;
-    for (const entry of entriesText) {
-        const match = entry.match(scaleDamageRegex);
-        if (!match) continue;
-
-        const tagContent = match[1];
-        const [baseDamage, levels, scalingDice] = tagContent.split('|');
-
-        const baseLevel = parseInt(levels.split('-')[0]);
-        const baseDiceMatch = baseDamage.match(/^(\d+)(d\d+)/);
-        const scaleDiceMatch = scalingDice.match(/^(\d+)/);
-
-        if (!baseDiceMatch || !scaleDiceMatch) continue;
-
-        const baseDiceCount = parseInt(baseDiceMatch[1]);
-        const diceFaces = baseDiceMatch[2];
-        const scaleDiceCount = parseInt(scaleDiceMatch[1]);
-
-        const scaling: { [key: string]: string } = {};
-
-        for (let lvl = baseLevel; lvl <= 9; lvl++) {
-            if (lvl === baseLevel) {
-                scaling[String(lvl)] = baseDamage;
-            } else {
-                const levelDiff = lvl - baseLevel;
-                const currentDiceCount = baseDiceCount + levelDiff * scaleDiceCount;
-                scaling[String(lvl)] = `${currentDiceCount}${diceFaces}`;
-            }
-        }
-
-        return [{ type: 'upcast', scaling }];
-    }
-
-    return null;
-}
-
-function getSpellMaterial(spell: any): string | null {
-    if (!spell.components.m) return null;
-    const material = spell.components.m;
-    if (typeof material === 'string') return material;
-    return material.text;
-}
-
 function getSpell(spell: any, fluffs: any[], sources: any): ParsedSpell {
     return {
         name: spell.name,
@@ -189,7 +113,7 @@ function getSpell(spell: any, fluffs: any[], sources: any): ParsedSpell {
         castingTime: parseCastingTime(spell.time, spell.meta),
         range: parseRange(spell.range),
         components: parseComponents(spell.components),
-        material: getSpellMaterial(spell),
+        material: parseMaterialComponents(spell.components),
         duration: parseDurationTime(spell.duration),
         url: getSpellsUrl(spell.name, spell.source),
         image: getSpellImage(fluffs, spell.name, spell.source),
@@ -204,7 +128,7 @@ function getSpell(spell: any, fluffs: any[], sources: any): ParsedSpell {
         conditionImmune: spell.conditionImmune ?? [],
         savingThrow: spell.savingThrow ?? [],
         affectsCreatureType: spell.affectsCreatureType ?? [],
-        scaledDamage: getSpellDamage(spell),
+        scaledDamage: parseSpellDamage(spell),
     };
 }
